新增通用的账号服务

+++ src/aliens/module/passport/db/model.go
@@ -13,14 +13,14 @@ const (
 
 //角色
 type DBUser struct {
-	ID       int32  `bson:"_id" gorm:"AUTO_INCREMENT"`             //用户id
+	ID       int64  `bson:"_id" gorm:"AUTO_INCREMENT"`             //用户id
 	Username string `bson:"username"  rorm:"uname"  unique:"true"` //用户名 渠道信息_渠道用户id存Username
 	Password string `bson:"password"  rorm:"pwd"`                  //加密的密码
 	Salt     string `bson:"salt"      rorm:"salt"`                 //加密的salt
 
 	ChannelUID string `bson:"cuid"    rorm:"cuid"`    //用户的渠道的渠道用户id
 	Channel    string `bson:"channel" rorm:"channel"` //用户的渠道信息 渠道用户id存Username
-	Avatar	   string `bson:"avatar"  rorm:"avatar"` //用户的头像地址
+	Avatar	   string `bson:"avatar"  rorm:"avatar"`  //用户的头像地址
 
 	Mobile string `bson:"mobile"` //用户电话
 	OpenID string `bson:"openid"` //微信OPENID 绑定微信填写
